create random agent

+++ cfg.py
@@ -1,8 +1,9 @@
 import utils
 import datetime
 import logging
-import json
+import pickle
 from pathlib import Path
+import json
 
 
 class CfgMaker():
@@ -21,8 +22,10 @@ class CfgMaker():
         
         self.conf_dir = Path(self.experiment_folder) / Path("config") 
         utils.make_dir(self.conf_dir)
-        self.conf_dir = self.conf_dir / 'config.json'
-        self.conf_dir.unlink()
+        self.conf_dir = self.conf_dir / 'config.pkl'
+        if self.conf_dir.exists():
+            self.conf_dir.unlink()
+        self.all_configs = dict()
 
 
     def make_cfg_logger(self):
@@ -36,35 +39,48 @@ class CfgMaker():
             'DEV_LEVEL'     : self.DEV_LEVEL,
             }
 
-        with open(str(self.conf_dir), 'a') as fp:
-            json.dump(self.cfg_logger, fp, indent=4)
+        self.all_configs['cfg_logger'] = self.cfg_logger
+        self.dump_cfg(self.all_configs)
+
         return self.cfg_logger
     
     def make_cfg_agent(self):
         self.cfg_agent = {
             'name'      : "cfg_agent",
         }
-        with open(str(self.conf_dir), 'a') as fp:
-            json.dump(self.cfg_agent, fp, indent=4)
+
+        self.all_configs['cfg_agent'] = self.cfg_agent
+        self.dump_cfg(self.cfg_agent)
+
         return self.cfg_agent
+
+    def update_cfg(self,old_dict,upd_dict):
+        self.all_configs[old_dict['name']].update(upd_dict)
+        self.dump_cfg(self.all_configs)
+        return self.cfg_agent
+
 
 
     def make_cfg_experiment(self):
         self.cfg_experiment = {
-            'name'      : "cfg_experiment",
+            'name'              : "cfg_experiment",
+            'max_allowed_steps' : 50,
         }
-        with open(str(self.conf_dir), 'a') as fp:
-            json.dump(self.cfg_experiment, fp, indent=4)
+
+        self.all_configs['cfg_experiment'] = self.cfg_experiment
+        self.dump_cfg(self.cfg_agent)
+
         return self.cfg_experiment
 
     def make_cfg_env(self):
         self.cfg_env = {
             'name'      :   'cfg_env',
-            'env_name'  :   'CartPole-v0',#'MountainCarContinuous-v0',
+            'env_name'  :   'Acrobot-v1',#'CartPole-v0',#'MountainCarContinuous-v0',
             'render'    :   True,
         }
-        with open(str(self.conf_dir), 'a') as fp:
-            json.dump(self.cfg_env, fp, indent=4)
+        self.all_configs['cfg_env'] = self.cfg_env
+        self.dump_cfg(self.cfg_env)
+
         return self.cfg_env
 
 
@@ -76,14 +92,26 @@ class CfgMaker():
             'dump_dir': str(dump_dir),
             'dump_file': 'dump.txt',
         }
-        with open(str(self.conf_dir), 'a') as fp:
-            json.dump(self.cfg_dumper, fp, indent=4)
+        self.all_configs['cfg_dumper'] = self.cfg_dumper
+        self.dump_cfg(self.cfg_env)
         return self.cfg_dumper
 
     def dump_cfg(self, cfg):
-        with open(str(self.conf_dir), 'a') as fp:
-            json.dump(cfg, fp, indent=4)
+        with open(str(self.conf_dir), 'wb') as fp:
+            pickle.dump(cfg, fp, protocol=pickle.HIGHEST_PROTOCOL)
+    
+    def load_cfg(self):
+        print(f"Loading configs from: {self.conf_dir}")
+        self.show_configs()
+        with open(str(self.conf_dir), 'rb') as handle:
+            self.all_configs = pickle.load(handle)
 
+    def show_configs(self):        
+        for conf_name, configs in self.all_configs.items():
+            try:
+                str_conf = json.dumps(configs, indent = 4)
+                print(f"\n {conf_name} -> {str_conf}")
+            except (TypeError, OverflowError):
+                print(f"\n {conf_name} -> {configs}")
 
-        
-
+
